write test for Lobster+combine.
